refactor: move fake classes to tests-utils/fake
